+++ src/cli/cli-commands/start-server.ts
@@ -13,8 +13,7 @@ import { startRemoteServer } from '../../remote-server/startRemoteServer';
 import { $provideFilesystemForNode } from '../../scrapers/_common/register/$provideFilesystemForNode';
 import { $provideScrapersForNode } from '../../scrapers/_common/register/$provideScrapersForNode';
 import { $provideScriptingForNode } from '../../scrapers/_common/register/$provideScriptingForNode';
-import type { number_port } from '../../types/typeAliases';
-import type { string_url } from '../../types/typeAliases';
+import type { number_port, string_url } from '../../types/typeAliases';
 import { suffixUrl } from '../../utils/normalization/suffixUrl';
 import { TODO_USE } from '../../utils/organization/TODO_USE';
 import { keepUnused } from '../../utils/organization/keepUnused';
@@ -98,10 +97,11 @@ export function $initializeStartServerCommand(program: Program) {
                 rootUrl = suffixUrl(url, '/books');
             }
 
-            let rootPath = '/';
+            const rootPath = '/';
 
-            if (url !== null) {
-                rootPath = url.pathname;
+            if (url !== null && url.pathname !== '/' && url.pathname !== '') {
+                console.error(colors.red(`URL of the server can not have path, but got "${url.pathname}"`));
+                process.exit(1);
             }
 
             // TODO: DRY [◽]